+++ admin/class-admin-asset-manager.php
@@ -418,7 +418,7 @@ class WPSEO_Admin_Asset_Manager {
 			'deps'    => [
 				'jquery',
 			],
-			'version' => '4.0.3',
+			'version' => '4.1.0-rc.0',
 		];
 		$scripts['select2-translations'] = [
 			'name'    => 'select2-translations',
@@ -427,7 +427,7 @@ class WPSEO_Admin_Asset_Manager {
 				'jquery',
 				self::PREFIX . 'select2-core',
 			],
-			'version' => '4.0.3',
+			'version' => '4.1.0-rc.0',
 		];
 
 		return $scripts;
@@ -552,7 +552,7 @@ class WPSEO_Admin_Asset_Manager {
 				'name'    => 'select2',
 				'src'     => 'select2/select2',
 				'suffix'  => '.min',
-				'version' => '4.0.1',
+				'version' => '4.1.0-rc.0',
 				'rtl'     => false,
 			],
 			[