+++ inc/options/class-wpseo-option-titles.php
@@ -621,7 +621,13 @@ class WPSEO_Option_Titles extends WPSEO_Option {
 					break;
 				case 'schema-article-type-':
 					if ( isset( $dirty[ $key ] ) && is_string( $dirty[ $key ] ) ) {
-						/** This filter is documented in inc/class-wpseo-meta.php */
+						/**
+						 * Filter: 'wpseo_schema_article_types' - Allow developers to filter the available article types.
+						 *
+						 * Make sure when you filter this to also filter `wpseo_schema_article_types_labels`.
+						 *
+						 * @api array $schema_article_types The available schema article types.
+						 */
 						if ( array_key_exists( $dirty[ $key ], apply_filters( 'wpseo_schema_article_types', Schema_Types::ARTICLE_TYPES ) ) ) {
 							$clean[ $key ] = $dirty[ $key ];
 						}
@@ -973,6 +979,14 @@ class WPSEO_Option_Titles extends WPSEO_Option {
 			'sc-smstar' => [
 				'option' => '&#8902;',
 				'label'  => __( 'Low asterisk', 'wordpress-seo' ),
+			],
+			'sc-pipe'   => [
+				'option' => '|',
+				'label'  => __( 'Vertical bar', 'wordpress-seo' ),
+			],
+			'sc-tilde'  => [
+				'option' => '~',
+				'label'  => __( 'Small tilde', 'wordpress-seo' ),
 			],
 			'sc-laquo'  => [
 				'option' => '&laquo;',