+++ packages/yoastseo/spec/worker/AnalysisWebWorkerSpec.js
@@ -1,6 +1,3 @@
-/**
- * @jest-environment jsdom
- */
 // External dependencies
 import { forEach, isArray, isNumber, isObject } from "lodash-es";
 import { getLogger } from "loglevel";