+++ packages/yoastseo/src/languageProcessing/researches/altTagCount.js
@@ -9,13 +9,14 @@ import { isEmpty } from "lodash-es";
  * Matches the alt-tags in the images found in the text.
  * Returns an object with the totals and different alt-tags.
  *
- * @param {Array} imageMatches  Array with all the matched images in the text
- * @param {Object} topicForms   The object with the keyphrase and the synonyms forms from the paper.
- * @param {string} locale       The locale used for transliteration.
+ * @param {Array}       imageMatches        	Array with all the matched images in the text
+ * @param {Object}      topicForms          	The object with the keyphrase and the synonyms forms from the paper.
+ * @param {string}      locale              	The locale used for transliteration.
+ * @param {function}    matchWordCustomHelper 	A language-specific helper function to match word in text.
  *
  * @returns {object} altProperties Object with all alt-tags that were found.
  */
-const matchAltProperties = function( imageMatches, topicForms, locale ) {
+const matchAltProperties = function( imageMatches, topicForms, locale, matchWordCustomHelper ) {
 	const altProperties = {
 		noAlt: 0,
 		withAlt: 0,
@@ -39,7 +40,7 @@ const matchAltProperties = function( imageMatches, topicForms, locale ) {
 		}
 
 		// If the keyword is matched in the alt tag
-		const keywordMatchedInAltTag = findTopicFormsInString( topicForms, alttag, true, locale );
+		const keywordMatchedInAltTag = findTopicFormsInString( topicForms, alttag, true, locale, matchWordCustomHelper );
 		if ( keywordMatchedInAltTag.percentWordMatches >= 50 ) {
 			altProperties.withAltKeyword++;
 			continue;
@@ -61,6 +62,7 @@ const matchAltProperties = function( imageMatches, topicForms, locale ) {
  */
 export default function altTagCount( paper, researcher ) {
 	const topicForms = researcher.getResearch( "morphology" );
+	const matchWordCustomHelper = researcher.getHelper( "matchWordCustomHelper" );
 
-	return matchAltProperties( imageInText( paper.getText() ), topicForms, paper.getLocale() );
+	return matchAltProperties( imageInText( paper.getText() ), topicForms, paper.getLocale(), matchWordCustomHelper );
 }