+++ src/integrations/watchers/indexable-author-watcher.php
@@ -83,6 +83,11 @@ class Indexable_Author_Watcher implements Integration_Interface {
 	 */
 	public function build_indexable( $user_id ) {
 		$indexable = $this->repository->find_by_id_and_type( $user_id, 'user', false );
-		$this->builder->build_for_id_and_type( $user_id, 'user', $indexable );
+		$indexable = $this->builder->build_for_id_and_type( $user_id, 'user', $indexable );
+
+		if ( $indexable ) {
+			$indexable->object_last_modified = \max( $indexable->object_last_modified, \current_time( 'mysql' ) );
+			$indexable->save();
+		}
 	}
 }